# src/swagger/openAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "E-Insurance App",
    "description": "This is an online insurance application. The application includes 4 user APIs and 3 service APIs.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
  "/scheme": {
      "post": {
          "tags": ["Scheme"],
          "summary": "Create a new scheme",
          "operationId": "createScheme",
          "requestBody": {
          "required": true,
          "content": {
              "application/json": {
              "schema": {
                  "$ref": "#/components/schemas/Scheme"
              }
              }
          }
          },
          "responses": {
          "201": {
              "description": "Scheme created successfully"
          },
          "400": {
              "description": "Bad request, validation failed"
          },
          "401": {
              "description": "Unauthorized, admin privileges required"
          },
          "500": {
              "description": "Internal server error"
          }
          }   
      },
      "get": {
          "tags": ["Scheme"],
          "summary": "Get all schemes with pagination",
          "operationId": "getAllSchemes",
          "parameters": [
          {
              "name": "page",
              "in": "query",
              "description": "Page number for pagination",
              "required": false,
              "schema": {
              "type": "integer",
              "default": 1
              }
          },
          {
              "name": "limit",
              "in": "query",
              "description": "Number of items per page for pagination",
              "required": false,
              "schema": {
              "type": "integer",
              "default": 10
              }
          }
          ],
          "responses": {
          "200": {
              "description": "A list of schemes with pagination"
          },
          "401": {
              "description": "Unauthorized, admin privileges required"
          },
          "500": {
              "description": "Internal server error"
          }
          }
      }
  },
...
  },
  "components": {
    "schemas": {
...
      "Scheme": {
        "type": "object",
        "properties": {
          "schemeName": { "type": "string" },
          "description": { "type": "string" },
          "planId": { "type": "string" },
          "eligibilityCriteria": { "type": "string" },
          "premium": { "type": "number", "format": "float" },
          "maturityPeriod": { "type": "number" },
          "coverage": { "type": "number" }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "name": "Authorization",
        "scheme": "bearer",
         "bearerFormat": "JWT",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}